+++ hamcrest-matchers/guava-hamcrest-matchers/src/test/java/it/ozimov/cirneco/hamcrest/guava/collect/IsMultimapKeyWithCollectionSizeTest.java
@@ -52,7 +52,7 @@ public class IsMultimapKeyWithCollectionSizeTest extends BaseMatcherTest {
         IsMultimapKeyWithCollectionSize.keyWithSize(comparison, -1);
 
         // Assert
-        fail("IllegalArgumentException expected");
+        fail("IllegalArgumentException comparison");
     }
 
     @Test