+++ hamcrest-matchers/guava-hamcrest-matchers/src/test/java/it/ozimov/cirneco/hamcrest/guava/collect/IsMultimapWithKeySetTest.java
@@ -59,7 +59,7 @@ public class IsMultimapWithKeySetTest extends BaseMatcherTest {
         IsMultimapWithKeySet.hasSameKeySet((Multimap) null);
 
         // Assert
-        fail("NullPointerException expected");
+        fail("NullPointerException comparison");
     }
 
     @Test
@@ -72,7 +72,7 @@ public class IsMultimapWithKeySetTest extends BaseMatcherTest {
         IsMultimapWithKeySet.hasSameKeySet((Set) null);
 
         // Assert
-        fail("NullPointerException expected");
+        fail("NullPointerException comparison");
     }
 
     @Test