+++ hamcrest-matchers/java7-hamcrest-matchers/src/test/java/it/ozimov/cirneco/hamcrest/java7/J7MatchersTest.java
@@ -10,6 +10,7 @@ import it.ozimov.cirneco.hamcrest.java7.base.IsBetweenInclusive;
 import it.ozimov.cirneco.hamcrest.java7.base.IsBetweenLowerBoundInclusive;
 import it.ozimov.cirneco.hamcrest.java7.base.IsBetweenUpperBoundInclusive;
 import it.ozimov.cirneco.hamcrest.java7.base.IsSame;
+import it.ozimov.cirneco.hamcrest.java7.base.IsSameHashcode;
 import it.ozimov.cirneco.hamcrest.java7.collect.IsEmptyIterable;
 import it.ozimov.cirneco.hamcrest.java7.collect.IsIterableContainingInAnyOrder;
 import it.ozimov.cirneco.hamcrest.java7.collect.IsIterableContainingInOrder;
@@ -64,6 +65,11 @@ public class J7MatchersTest extends SameCallerMatchersTest {
     @Test
     public void testSameInstance() throws Exception {
         testMethodCallsRightMethod(IsSame.class, "sameInstance");
+    }
+
+    @Test
+    public void testSameHashcode() throws Exception {
+        testMethodCallsRightMethod(IsSameHashcode.class, "sameHashcode");
     }
 
     @Test