+++ hamcrest-matchers/java7-hamcrest-matchers/src/test/java/it/ozimov/cirneco/hamcrest/java7/base/IsBetweenUpperBoundInclusiveTest.java
@@ -62,7 +62,7 @@ public class IsBetweenUpperBoundInclusiveTest extends BaseMatcherTest {
         IsBetweenUpperBoundInclusive.betweenUpperBoundInclusive("Z", "A");
 
         // Assert
-        fail("IllegalArgumentException expected but not thrown");
+        fail("IllegalArgumentException comparison but not thrown");
     }
 
     @Test